backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
base58==2.1.1
secp256k1==0.14.0
mnemonic==0.21